refactor(uanodeset-rdf): separate collecting a model from mapping it
The mapper read an IAddressSpace directly, which tied it to a source that
answers synchronously and at random. A session answers asynchronously and in
batches, and rewriting the mapper around await would have turned four passes
over the nodes into thousands of round trips.

So a collector now fills an RdfModel and the mapper reads only that, staying
synchronous. RdfModel is deliberately close to what OPC UA can actually say
about a node, since the collector this makes room for has only Browse and Read.

Foreign nodes are converted lazily: the mapper names supertypes and type
definitions from imported models, and walking all of Opc.Ua to find the handful
it asks for would cost more than it saves.

No behaviour change, which is the point of the commit being on its own: the
export of DI hashes to what it hashed before.
